# reland the 1x2 unrolling optimization