Summary:
took Ilya's idea in D38803011 and D38543825 (D38543825 is a hacky unrolling which only deals with k%4 == 2),  implement it properly into FbgemmFP16UKernelsAvx2.cc.

before:
```

                      MKL_FP32 m =     1 n =    16 k =   162 Gflops =   9.5837 GBytes =  10.8999
                         FBP_t m =     1 n =    16 k =   162 Gflops =   8.7079 GBytes =   9.9039

                      MKL_FP32 m =     1 n =    16 k =   163 Gflops =   9.8374 GBytes =  11.1878
                         FBP_t m =     1 n =    16 k =   163 Gflops =  13.3576 GBytes =  15.1912

                      MKL_FP32 m =     1 n =    16 k =   164 Gflops =   9.8970 GBytes =  11.2549
                         FBP_t m =     1 n =    16 k =   164 Gflops =  13.3830 GBytes =  15.2191

                      MKL_FP32 m =     1 n =    16 k =   165 Gflops =   9.8797 GBytes =  11.2344
                         FBP_t m =     1 n =    16 k =   165 Gflops =  12.2774 GBytes =  13.9608

                      MKL_FP32 m =     1 n =    16 k =   166 Gflops =   9.3144 GBytes =  10.5909
                         FBP_t m =     1 n =    16 k =   166 Gflops =  13.4026 GBytes =  15.2394

                      MKL_FP32 m =     1 n =    16 k =   167 Gflops =   9.8592 GBytes =  11.2097
                         FBP_t m =     1 n =    16 k =   167 Gflops =  13.4705 GBytes =  15.3156

                      MKL_FP32 m =     1 n =    16 k =   168 Gflops =   8.1439 GBytes =   9.2588
                         FBP_t m =     1 n =    16 k =   168 Gflops =  13.4329 GBytes =  15.2720

                      MKL_FP32 m =     1 n =    16 k =   169 Gflops =   8.9921 GBytes =  10.2225
                         FBP_t m =     1 n =    16 k =   169 Gflops =  13.3934 GBytes =  15.2261
```
after:
```

                      MKL_FP32 m =     1 n =    16 k =   162 Gflops =   8.9189 GBytes =  10.1438
                         FBP_t m =     1 n =    16 k =   162 Gflops =  19.2140 GBytes =  21.8530

                      MKL_FP32 m =     1 n =    16 k =   163 Gflops =   9.1963 GBytes =  10.4587
                         FBP_t m =     1 n =    16 k =   163 Gflops =  18.1100 GBytes =  20.5960

                      MKL_FP32 m =     1 n =    16 k =   164 Gflops =   7.9533 GBytes =   9.0444
                         FBP_t m =     1 n =    16 k =   164 Gflops =  19.7641 GBytes =  22.4756

                      MKL_FP32 m =     1 n =    16 k =   165 Gflops =   8.7126 GBytes =   9.9072
                         FBP_t m =     1 n =    16 k =   165 Gflops =  18.1491 GBytes =  20.6378

                      MKL_FP32 m =     1 n =    16 k =   166 Gflops =   8.3245 GBytes =   9.4653
                         FBP_t m =     1 n =    16 k =   166 Gflops =  18.8134 GBytes =  21.3917

                      MKL_FP32 m =     1 n =    16 k =   167 Gflops =   8.9353 GBytes =  10.1592
                         FBP_t m =     1 n =    16 k =   167 Gflops =  19.5768 GBytes =  22.2584

                      MKL_FP32 m =     1 n =    16 k =   168 Gflops =   9.2355 GBytes =  10.4999
                         FBP_t m =     1 n =    16 k =   168 Gflops =  19.1443 GBytes =  21.7652

                      MKL_FP32 m =     1 n =    16 k =   169 Gflops =   9.0509 GBytes =  10.2894
                         FBP_t m =     1 n =    16 k =   169 Gflops =  20.5138 GBytes =  23.3207

Differential Revision: D39214267